@@ -1,9 +1,13 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import OneBacklog from './OneBacklog.vue'
 import axios from 'axios'
 import router from './router'
 import store from './store/store'
-import './fa.config'
+import mitt from 'mitt'
+import FontAwesomeIcon from './fa.config'
+// import the global css, see https://stackoverflow.com/questions/39438094/best-way-to-have-global-css-in-vuejs
+import '@/css/onebacklog.scss'
+import '@/css/onebacklog.css'
 
 axios.defaults.baseURL = import.meta.env.VITE_API_URL
 axios.defaults.withCredentials = true
@@ -14,17 +18,15 @@ const reqInterceptor = axios.interceptors.request.use(config => {
 const resInterceptor = axios.interceptors.response.use(res => {
 	return res
 })
-
 axios.interceptors.request.eject(reqInterceptor)
 axios.interceptors.response.eject(resInterceptor)
 
-export const eventBus = new Vue()
-// import the global css, see https://stackoverflow.com/questions/39438094/best-way-to-have-global-css-in-vuejs
-import '@/css/onebacklog.scss'
-import '@/css/onebacklog.css'
-new Vue({
-	el: '#onebacklog',
-	router,
-	store,
-	render: h => h(OneBacklog)
-})
+const eventBus = mitt()
+
+const app = createApp(OneBacklog)
+app.use(router)
+app.use(store)
+app.component("font-awesome-icon", FontAwesomeIcon)
+app.config.globalProperties.eventBus = eventBus
+
+app.mount('#onebacklog')
