@@ -2,7 +2,7 @@
   <div>
     <b-navbar toggleable="md" type="dark" variant="dark">
       <b-navbar-toggle target="nav_collapse"></b-navbar-toggle>
-      <b-img class="logo" :src="require('../../assets/logo.png')" alt="OneBacklog logo" />
+      <b-img class="logo" :src="logo" alt="OneBacklog logo" />
       <b-navbar-brand to="../../rel-notes">{{ appVersion }}</b-navbar-brand>
       <b-collapse is-nav id="nav_collapse">
         <b-navbar-nav>
