@@ -4,7 +4,6 @@
 import { SEV, LEVEL } from '../../../constants.js'
 import { collapseNode, expandNode } from '../../../common_functions.js'
 import commonView from '../common_view.js'
-import { eventBus } from '../../../main.js'
 
 // in px; used to calculate the node closest to the current cursor position
 const edgeSize = 6
@@ -101,7 +100,7 @@ const methods = {
 
 	// trigger the context component via the eventbus if the node is selectable
 	emitNodeContextmenu(node) {
-		if (node.isSelectable) eventBus.$emit('context-menu', node)
+		if (node.isSelectable) this.eventBus.$emit('context-menu', node)
 	},
 
 	emitSelect(fromContextMenu) {
